Moved functions to utilils.js
